backend/pyproject.toml
[project]
name = "drawerlabels"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlmodel>=0.0.21",
    "qrcode[pil]>=7.4",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
